examples/authorization-for-rag/langchain-examples/pyproject.toml
[tool.poetry]
name = "langchain-examples"
version = "0.1.0"
description = ""
authors = ["Auth0 <support@auth0.com>"]
readme = "README.md"
packages = [{ include = "langchain_rag" }]

[tool.poetry.scripts]
fga_init = "scripts.fga_init:main"
start = "langchain_rag.main:main"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.11"
langchain-openai = "^0.2.12"
faiss-cpu = "^1.9.0.post1"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.11"
auth0-ai-langchain = { path = "../../../packages/auth0-ai-langchain", develop = true }
openfga-sdk = "^0.9.0"
termcolor = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
